refactor: rename Navbar again
